21 12 06 01
clock building and old house finished making all 9 buildings. We will end the building design for now on because adding more buildings are making the game a little more easy. Making more buildings lead to flat ground of the buildings. We will work on present designs, sounds, IAP systems, and bugs if found during the making process of this game. Have a great night and see you guys tomorrow!